Created Apriori module
