--- MagicMine/Algorithms/cluster.py
@@ -1,1 +1,0 @@
-#Matriz de correlaciones
