--- MagicMine/Templates/aprioriProcess.html
@@ -1,76 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Tabla de frecuencias</title>
-    <style>
-        table {
-            border-collapse: collapse;
-            width: 50%; /* Reducir el ancho de la tabla al 50% */
-        }
-
-        th, td {
-            border: 1px solid #ddd;
-            padding: 4px; /* Reducir el relleno de las celdas de la tabla */
-            text-align: left;
-        }
-
-        th {
-            background-color: #f2f2f2;
-        }
-
-        .small-image {
-            max-width: 50%; /* Reducir el ancho máximo de la imagen al 50% */
-            height: auto; /* Ajustar automáticamente la altura de la imagen */
-        }
-
-        .button-container {
-            text-align: center;
-            margin-top: 20px;
-        }
-
-        .button-container a {
-            display: inline-block;
-            padding: 10px 20px;
-            background-color: #333;
-            color: #fff;
-            text-decoration: none;
-            border-radius: 4px;
-        }
-    </style>
-</head>
-<body>
-    <h1>Resultado</h1>
-    <div>
-        {{ dataframe|safe }}
-    </div>
-    <h1>Gráfico de frecuencias</h1>
-    <img src="{{ url_for('static', filename='chart.png') }}" alt="Gráfico de Barras" class="small-image">
-    <div>
-        <h1>Tabla de Reglas de Asociación</h1>
-        <table>
-            <thead>
-                <tr>
-                    <th>Regla de Asociación</th>
-                    <th>Soporte</th>
-                    <th>Confianza</th>
-                    <th>Elevación</th>
-                </tr>
-            </thead>
-            <tbody>
-                {% for rule in enumerated_rules %}
-                    <tr>
-                        <td>{{ rule[0] }}</td>
-                        <td>{{ rule[1] }}</td>
-                        <td>{{ rule[2] }}</td>
-                        <td>{{ rule[3] }}</td>
-                    </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-    </div>    
-    <div class="button-container">
-        <a href="/apriori">Regresar</a>
-    </div>
-    
-</body>
-</html>
