--- MagicMine/Templates/menu.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="UTF-8">
-  <title>Menú</title>
-  <style>
-    ul {
-      list-style-type: none;
-      margin: 0;
-      padding: 0;
-      background-color: #f1f1f1;
-    }
-
-    li {
-      display: inline-block;
-    }
-
-    li a {
-      display: block;
-      padding: 10px 20px;
-      text-decoration: none;
-      color: #333;
-      font-weight: bold;
-    }
-
-    li a:hover {
-      background-color: #ccc;
-    }
-  </style>
-</head>
-<body>
-  <ul>
-    <li><a href="apriori">Apriori</a></li>
-    <li><a href="metricas">Metricas de distancia</a></li>
-    <li><a href="clustering">Clustering</a></li>
-    <li><a href="upload">Subir archivo</a></li>
-  </ul>
-</body>
-</html>
