+++ MagicMine/src/templates/home.html
@@ -1,0 +1,61 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <meta charset="UTF-8">
+  <title>MagicMine - Página de Inicio</title>
+  <style>
+    body {
+      background-color: #f2f2f2;
+      font-family: Arial, sans-serif;
+      margin: 0;
+      padding: 0;
+    }
+    
+    .container {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      justify-content: center;
+      height: 100vh;
+      text-align: center;
+    }
+    
+    .logo {
+      margin-bottom: 20px;
+    }
+    
+    h1 {
+      font-size: 48px;
+      color: #333;
+      margin-bottom: 20px;
+    }
+    
+    p {
+      font-size: 24px;
+      color: #777;
+      margin-bottom: 40px;
+    }
+    
+    a {
+      text-decoration: none;
+      background-color: #333;
+      color: #fff;
+      padding: 10px 20px;
+      font-size: 20px;
+      border-radius: 4px;
+    }
+    
+    a:hover {
+      background-color: #555;
+    }
+  </style>
+</head>
+<body>
+  <div class="container">
+    <img class="logo" src="/src/logo.jpg" alt="MagicMine Logo">
+    <h1>MagicMine</h1>
+    <p>Bienvenido a MagicMine, el lugar donde la magia se hace con algoritmos.</p>
+    <a href="menu">Comenzar</a>
+  </div>
+</body>
+</html>
