+++ MagicMine/src/templates/metricas.html
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Metricas de distancia</title>
+</head>
+<body>
+    <h1>Seleccionar archivo CSV</h1>
+    <form id="metricas-form" action="/metricas/process" method="POST">
+        <select name="file">
+            {% for csv_file in csv_files %}
+            <option value="{{ csv_file }}">{{ csv_file }}</option>
+            {% endfor %}
+        </select>
+        <br>
+        <label for="option-select">Elige el tipo de distancia:</label>
+        <select name="option" id="option-select">
+            <option value="Euclidiana">Euclidiana</option>
+            <option value="Chebyshev">Chebyshev</option>
+            <option value="Manhattan">Manhattan</option>
+            <option value="Minkowski">Minkowski</option>
+        </select>
+        <br>
+        <input type="submit" value="Procesar">
+    </form>
+    <form action="/menu" method="GET">
+        <input type="submit" value="Regresar">
+    </form>
+
+    <script>
+        document.getElementById("metricas-form").addEventListener("submit", function(e) {
+            
+        });
+    </script>
+</body>
+</html>
