+++ MagicMine/src/templates/metricasEuclidiana.html
@@ -1,0 +1,56 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Tabla de frecuencias</title>
+    <style>
+        table {
+            border-collapse: collapse;
+            width: 50%; /* Reducir el ancho de la tabla al 50% */
+        }
+
+        th, td {
+            border: 1px solid #ddd;
+            padding: 4px; /* Reducir el relleno de las celdas de la tabla */
+            text-align: left;
+        }
+
+        th {
+            background-color: #f2f2f2;
+        }
+
+        .small-image {
+            max-width: 50%; /* Reducir el ancho máximo de la imagen al 50% */
+            height: auto; /* Ajustar automáticamente la altura de la imagen */
+        }
+
+        .button-container {
+            text-align: center;
+            margin-top: 20px;
+        }
+
+        .button-container a {
+            display: inline-block;
+            padding: 10px 20px;
+            background-color: #333;
+            color: #fff;
+            text-decoration: none;
+            border-radius: 4px;
+        }
+    </style>
+</head>
+<body>
+    <h1>Resultado</h1>
+    <div>
+        <h1>Matriz estandarizada</h1>>
+        {{ euclidianSample|safe }}
+    </div>    
+    <div>
+        <h1>Matriz euclidiana</h1>
+        {{ MEuclidiana | safe}}
+    </div>
+    <div class="button-container">
+        <a href="/metricas">Regresar</a>
+    </div>
+    
+</body>
+</html>
