+++ MagicMine/src/templates/upload.html
@@ -1,0 +1,81 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Subir Archivo CSV</title>
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            margin: 20px;
+        }
+        h1 {
+            text-align: center;
+            margin-bottom: 20px;
+        }
+        .upload-form {
+            max-width: 400px;
+            margin: 0 auto;
+        }
+        .upload-form label {
+            display: block;
+            margin-bottom: 10px;
+        }
+        .upload-form input[type="file"] {
+            display: block;
+            margin-bottom: 20px;
+        }
+        .upload-form input[type="submit"] {
+            background-color: #4CAF50;
+            color: white;
+            padding: 10px 20px;
+            border: none;
+            cursor: pointer;
+        }
+        .message {
+            text-align: center;
+            margin-top: 20px;
+        }
+        .success-message {
+            color: #4CAF50;
+        }
+        .error-message {
+            color: #FF0000;
+        }
+        .return-button {
+            text-align: center;
+            margin-top: 20px;
+        }
+        .return-button a {
+            background-color: #333;
+            color: #fff;
+            padding: 10px 20px;
+            text-decoration: none;
+        }
+        .return-button a:hover {
+            background-color: #555;
+        }
+    </style>
+</head>
+<body>
+    <h1>Subir Archivo CSV</h1>
+    <div class="upload-form">
+        {% if success_message %}
+            <div class="message success-message">
+                {{ success_message }}
+            </div>
+        {% endif %}
+        {% if error_message %}
+            <div class="message error-message">
+                {{ error_message }}
+            </div>
+        {% endif %}
+        <form action="/upload" method="POST" enctype="multipart/form-data">
+            <label for="file">Seleccionar archivo:</label>
+            <input type="file" id="file" name="file" accept=".csv" required>
+            <input type="submit" value="Subir">
+        </form>
+    </div>
+    <div class="return-button">
+        <a href="menu">Regresar</a>
+    </div>
+</body>
+</html>
